Finished implementing CRUD
