+++ src/delete.js
@@ -2,7 +2,6 @@ export const handleDelete = (deleteDivElement) => {
   const deleteKey = deleteDivElement.getAttribute('data-key');
   const deleteTdElement = deleteDivElement.closest('td');
   deleteTdElement.removeChild(deleteDivElement);
-
   const deleteStoredDataString = localStorage.getItem('userData');
   const deleteStoredData = deleteStoredDataString ? JSON.parse(deleteStoredDataString) : {};
   delete deleteStoredData[deleteKey];
