+++ src/display.js
@@ -1,23 +1,31 @@
-const displayStoredData = (tdElement, data) => {
+const renderData = (tdElement, data) => {
+  let content = '';
+  Object.keys(data).forEach((key) => {
+    if (tdElement.className.startsWith(key.substring(0, 3))) {
+      content += `
+        <div data-key="${key}">
+          <p>Task: ${data[key].task}</p>
+          <p>Date: ${data[key].date}</p>
+          <p>From: ${data[key].fromTime}</p>
+          <p>To: ${data[key].toTime}</p>
+          <button class="edit">Edit</button>
+          <button class="delete">Delete</button>
+        </div>
+      `;
+    }
+  });
+  tdElement.innerHTML += content;
+};
+
+const displayStoredData = (tdElements, data) => {
   if (data) {
-    tdElement.forEach((cells) => {
-      const content = '';
-      Object.keys(data).forEach((key) => {
-        if (cells.className.startsWith(key.substring(0, 3))) {
-          cells.innerHTML += `
-          <div data-key="${key}">
-            <p>Task: ${data[key].task}</p>
-            <p>Date: ${data[key].date}</p>
-            <p>From: ${data[key].fromTime}</p>
-            <p>To: ${data[key].toTime}</p>
-            <button class="edit">Edit</button>
-            <button class="delete">Delete</button>
-          </div>
-        `;
-        }
+    if (tdElements instanceof NodeList) {
+      tdElements.forEach((tdElement) => {
+        renderData(tdElement, data);
       });
-      tdElement.innerHTML = content;
-    });
+    } else {
+      renderData(tdElements, data);
+    }
   }
 };
 export default displayStoredData;
